roadsafety/requirements.txt
astroid==2.1.0
autopep8==1.4.3
Django==2.1.5
djangorestframework==3.9.1
gitdb2==2.0.5
GitPython==2.1.11
isort==4.3.4
lazy-object-proxy==1.3.1
lint==1.2.1
mccabe==0.6.1
mysqlclient==1.4.1
pycodestyle==2.5.0
pylint==2.2.2
pytz==2018.9
six==1.12.0
smmap2==2.0.5
typed-ast==1.2.0
wrapt==1.11.1
